applying a new project structure
