+++ mobile/__tests__/faq-test.tsx
@@ -1,4 +1,4 @@
-import FAQ from "../src/screens/faq/ui/faq-screen";
+import FAQ from "../src/screens/faq";
 import React from "react";
 import renderer from "react-test-renderer";
