+++ mobile/src/assets/png/adaptive-icon.png
