+++ mobile/src/assets/ttf/spacemono-regular.ttf
