--- mobile/src/components/contribute/index.ts
@@ -1,2 +1,0 @@
-// export contribute components
-export { default as CardItem } from "./card-item";
