--- mobile/src/components/shared/drawer-content.tsx
@@ -1,28 +1,0 @@
-import {
-  DrawerContentComponentProps,
-  DrawerContentOptions,
-  DrawerContentScrollView,
-  DrawerItemList,
-} from "@react-navigation/drawer";
-import { Image, SafeAreaView, View } from "react-native";
-import React, { FC } from "react";
-import { drawerStyles } from "../../styles";
-
-const DrawerContent: FC<DrawerContentComponentProps<DrawerContentOptions>> = (props) => {
-  return (
-    <DrawerContentScrollView {...props}>
-      <SafeAreaView>
-        {/* dzcode.io logo */}
-        <View>
-          <Image
-            source={require("../../assets/images/png/logo.png")}
-            style={drawerStyles.logoView}
-          />
-        </View>
-        {/* Drawer items */}
-        <DrawerItemList {...props} />
-      </SafeAreaView>
-    </DrawerContentScrollView>
-  );
-};
-export default DrawerContent;
