--- mobile/src/components/shared/index.ts
@@ -1,4 +1,0 @@
-// export shared components
-export { default as AppBar } from "./app-bar";
-export { default as DrawerContent } from "./drawer-content";
-export { default as DZCodeLoading } from "./dzcode-loading";
