--- mobile/src/screens/articles/articles-list/functions/index.ts
@@ -1,2 +1,0 @@
-// write your Articles functions here
-export {};
