+++ mobile/src/screens/contribute/styles.ts
@@ -1,0 +1,11 @@
+import { MEDIUM_MARGIN_SIZE } from "../../utils/constants";
+import { StyleSheet } from "react-native";
+
+export const contributeStyles = StyleSheet.create({
+  dzcodeLoading: {
+    position: "absolute",
+  },
+  listView: {
+    marginVertical: MEDIUM_MARGIN_SIZE,
+  },
+});
