--- mobile/src/screens/faq/functions/index.ts
@@ -1,2 +1,0 @@
-// write your FAQ functions here
-export {};
