--- mobile/src/screens/faq/ui/faq-screen.tsx
@@ -1,18 +1,0 @@
-import {} from "../functions";
-import React, { FC } from "react";
-import { Text } from "react-native-paper";
-import { View } from "react-native";
-import { globalStyles } from "../../../styles";
-
-const FAQScreen: FC = () => {
-  return (
-    // main view
-    <View style={globalStyles.mainView}>
-      {/* center view */}
-      <View style={globalStyles.centerView}>
-        <Text style={globalStyles.titleText}>FAQ UI created!</Text>
-      </View>
-    </View>
-  );
-};
-export default FAQScreen;
