--- mobile/src/screens/home/functions/index.ts
@@ -1,2 +1,0 @@
-// write your Home functions here
-export {};
