--- mobile/src/screens/learn/functions/index.ts
@@ -1,2 +1,0 @@
-// write your Learn functions here
-export {};
