--- mobile/src/screens/projects/functions/index.ts
@@ -1,2 +1,0 @@
-// write your Projects functions here
-export {};
