+++ mobile/src/styles/index.ts
@@ -1,7 +1,2 @@
 export { default as Colors } from "./colors";
-
 export { globalStyles } from "./global-styles";
-export { drawerStyles } from "./drawer-styles";
-export { cardStyles } from "./card-styles";
-export { dzcodeLoadingStyles } from "./dzcode-loading-styles";
-export { contributeStyles } from "./contribute-styles";
